refactor: Reduce layers in proxy
