--- app/proxy/proxy.py
@@ -1,45 +1,0 @@
-import logging
-from abc import ABC, abstractmethod
-from dataclasses import dataclass
-
-from app.errors import ModelExecutionError
-from app.models import MessageModel
-from app.proxy.policy import PolicyInterface
-
-log = logging.getLogger(__name__)
-
-
-class ProxyInterface(ABC):
-    @abstractmethod
-    async def valid_message(self, message: str) -> bool:
-        raise NotImplementedError
-
-
-@dataclass
-class Proxy(ProxyInterface):
-    policy: PolicyInterface
-
-    async def valid_message(self, message: str) -> bool:
-        """Validate if the message is allowed to be processed.
-        If the action is deny, will return false.
-        If the action is warn, will return false. And notify to alert service.
-        If the action is allow, will return true.
-        """
-        try:
-            policy_action = await self.policy.decide_policy_action(message)
-        except ModelExecutionError as e:
-            log.error(
-                "Model execution error on deciding policy action: Maybe prohibited message received from user or LLM: %s",
-                e,
-            )
-            return False
-        if policy_action == "deny":
-            return False
-        elif policy_action == "warn":
-            await self.policy.notify_external_service(message)
-            return False
-        elif policy_action == "allow":
-            return True
-        else:
-            log.warning("Unknown policy action: %s", policy_action)
-            return False
